Rename files in naming convention
