+++ src/colormap_container.ts
