+++ src/fixed_res_buffer.ts
